Extract globals domain
Make sure the global static values are centralized for reuse across
the application.
