--- client/globals.d.ts
@@ -1,4 +1,0 @@
-interface Window {
-  __GISTPEN_I18N__?: { [key: string]: string };
-  __webpack_public_path__?: string;
-}
