+++ client/globals/state.ts
