+++ client/reducers/index.ts
@@ -10,4 +10,3 @@ export * from './search';
 export * from './jobs';
 export * from './runs';
 export * from './messages';
-export * from './globals';
